Add 'ironscheme' command for IronScheme
This allows easy starting of a REPL from the shell. The wrapper needed
to start the REPL appears to be the exact same one as the one needed
for the scheme-script shebang interpreter. Since 'ironscheme' can be
used to start ironscheme for any purpose and 'scheme-script' is only
intended for one purpose, 'ironscheme' is more general and I made
'scheme-script' a symlink to it.